+++ src/navigation/Main/MainNavigator.tsx
@@ -1,7 +1,7 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import React from 'react';
 import { MAIN_ROUTES } from '~navigation/Main/mainTypes';
-import { EntityScreen, HomeScreen } from '~screens';
+import { EntityScreen, HomeScreen, SearchScreen } from '~screens';
 import { HomeHeader } from '~components';
 
 const Stack = createNativeStackNavigator();
@@ -16,6 +16,7 @@ export const MainNavigator = () => {
           options={{ headerShown: true, header: (props) => <HomeHeader {...props} /> }}
         />
         <Stack.Screen name={MAIN_ROUTES.ENTITY} component={EntityScreen} />
+        <Stack.Screen name={MAIN_ROUTES.SEARCH} component={SearchScreen} />
       </Stack.Group>
     </Stack.Navigator>
   );